+++ middleware/logger/config.go
@@ -48,20 +48,11 @@ type Config struct {
 	// Defaults to false.
 	TraceRoute bool
 
-	// Columns will display the logs as a formatted columns-rows text (bool).
-	// If custom `LogFunc` has been provided then this field is useless and users should
-	// use the `Columinize` function of the logger to get the output result as columns.
-	//
-	// Defaults to false.
-	Columns bool
-
 	// MessageContextKeys if not empty,
 	// the middleware will try to fetch
 	// the contents with `ctx.Values().Get(MessageContextKey)`
 	// and if available then these contents will be
 	// appended as part of the logs (with `%v`, in order to be able to set a struct too),
-	// if Columns field was set to true then
-	// a new column will be added named 'Message'.
 	//
 	// Defaults to empty.
 	MessageContextKeys []string
@@ -71,8 +62,6 @@ type Config struct {
 	// the contents with `ctx.Values().Get(MessageHeaderKey)`
 	// and if available then these contents will be
 	// appended as part of the logs (with `%v`, in order to be able to set a struct too),
-	// if Columns field was set to true then
-	// a new column will be added named 'HeaderMessage'.
 	//
 	// Defaults to empty.
 	MessageHeaderKeys []string
@@ -93,7 +82,7 @@ type Config struct {
 }
 
 // DefaultConfig returns a default config
-// that have all boolean fields to true except `Columns`,
+// that have all boolean fields to true,
 // all strings are empty,
 // LogFunc and Skippers to nil as well.
 func DefaultConfig() Config {
@@ -105,7 +94,6 @@ func DefaultConfig() Config {
 		PathAfterHandler: false,
 		Query:            false,
 		TraceRoute:       false,
-		Columns:          false,
 		LogFunc:          nil,
 		LogFuncCtx:       nil,
 		Skippers:         nil,